+++ go-server/frontend/src/tab_functions/inner_console_tab.js
@@ -59,8 +59,9 @@ export var v_createConsoleTabFunction = function () {
 	// Creating console tab in the inner tab list
 	var v_tab = v_connTabControl.selectedTab.tag.tabControl.createTab({
 		p_icon: '<i class="fas fa-terminal icon-tab-title"></i>',
-		p_name:
-			'<span> Console</span><span id="tab_loading" style="display:none;"><i class="tab-icon node-spin"></i></span><i title="" id="tab_check" style="display: none;" class="fas fa-check-circle tab-icon icon-check"></i></span>',
+		p_name: "Console",
+		p_status:
+			'<span id="tab_loading" style="display:none;"><i class="tab-icon node-spin"></i></span><i title="" id="tab_check" style="display: none;" class="fas fa-check-circle tab-icon icon-check"></i>',
 		p_selectFunction: function () {
 			if (this.tag != null) {
 				this.tag.resize();